+++ rustrade-execution/src/lib.rs
@@ -75,7 +75,9 @@ pub mod client;
 pub mod error;
 pub mod exchange;
 pub mod fee;
-pub use fee::{FeeModel, FeeModelConfig, PerContractFeeModel, PercentageFeeModel, ZeroFeeModel};
+pub use fee::{
+    FeeModel, FeeModelConfig, Liquidity, PerContractFeeModel, PercentageFeeModel, ZeroFeeModel,
+};
 pub mod fill;
 pub use fill::{BidAskFillModel, FillModel, LastPriceFillModel, MidpointFillModel, SimFillConfig};
 pub mod indexer;